chore: Refactor commons
