+++ src/pages/table-select-filter/index.jsx
@@ -3,21 +3,9 @@
 import React, { useLayoutEffect, useRef, useState } from 'react';
 import { createRoot } from 'react-dom/client';
 import { useCollection } from '@cloudscape-design/collection-hooks';
-import {
-  COLUMN_DEFINITIONS,
-  VISIBLE_CONTENT_OPTIONS,
-  PAGE_SIZE_OPTIONS,
-  SEARCHABLE_COLUMNS,
-} from './table-select-filter-config';
-import {
-  Button,
-  CollectionPreferences,
-  Input,
-  Pagination,
-  SpaceBetween,
-  Select,
-  Table,
-} from '@cloudscape-design/components';
+import { COLUMN_DEFINITIONS, SEARCHABLE_COLUMNS } from './table-select-filter-config';
+import { Preferences } from '../commons/table-config';
+import { Button, Input, Pagination, SpaceBetween, Select, Table } from '@cloudscape-design/components';
 import { Navigation, Breadcrumbs, ToolsContent } from './table-select-filter-components';
 import '../../styles/table-select.scss';
 import DATA from '../../resources/instances';
@@ -185,35 +173,7 @@ function TableSelectFilter({ loadHelpPanelContent }) {
         </div>
       }
       pagination={<Pagination {...paginationProps} ariaLabels={paginationAriaLabels} />}
-      preferences={
-        <CollectionPreferences
-          title="Preferences"
-          confirmLabel="Confirm"
-          cancelLabel="Cancel"
-          preferences={preferences}
-          onConfirm={({ detail }) => setPreferences(detail)}
-          pageSizePreference={{
-            title: 'Page size',
-            options: PAGE_SIZE_OPTIONS,
-          }}
-          wrapLinesPreference={{
-            label: 'Wrap lines',
-            description: 'Check to see all the text and wrap the lines',
-          }}
-          stripedRowsPreference={{
-            label: 'Striped rows',
-            description: 'Check to add alternating shaded rows',
-          }}
-          contentDensityPreference={{
-            label: 'Compact mode',
-            description: 'Check to display table content in a denser, more compact mode',
-          }}
-          visibleContentPreference={{
-            title: 'Select visible columns',
-            options: VISIBLE_CONTENT_OPTIONS,
-          }}
-        />
-      }
+      preferences={<Preferences preferences={preferences} setPreferences={setPreferences} />}
     />
   );
 }
